move one test to the new test file
